Disable CSRF on info endpoint for testing
The info endpoint does not work in prod/pre-prod/dev via the Ingress,
which adds a prefix to all endpoints.

It also fails to work due to csrf protection, which causes our tests
to fail.

Disable CSRF on the info endpoint for tests only.

We have a ticket to fix the /info and /health endpoints in other
environments.